Feign client added to note service.

+++ sanotes-note-service/src/main/java/sanotesnoteservice/service/impl/NoteServiceImpl.java
@@ -3,13 +3,15 @@ package sanotesnoteservice.service.impl;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeMap;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import sanotesnoteservice.client.NoteBookClient;
 import sanotesnoteservice.exeption.ResourceNotFoundException;
+import sanotesnoteservice.exeption.UnauthorizedException;
 import sanotesnoteservice.model.NoteContainerModel;
 import sanotesnoteservice.model.NoteModel;
 import sanotesnoteservice.model.NoteVersionModel;
 import sanotesnoteservice.payload.ApiResponse;
+import sanotesnoteservice.payload.BooleanResponse;
 import sanotesnoteservice.payload.ByIdRequest;
 import sanotesnoteservice.repository.NoteContainerRepository;
 import sanotesnoteservice.repository.NoteRepository;
@@ -31,15 +33,19 @@ public class NoteServiceImpl implements NoteService {
 
     private final ModelMapper modelMapper;
 
+    private final NoteBookClient noteBookClient;
+
     private static final String USER_DONT_HAVE_PERMISSION = "User don't have permission for this request";
 
 
-    public NoteContainerModel saveNote(NoteContainerModel noteContainerModel) {
+    public NoteContainerModel saveNote(NoteContainerModel noteContainerModel, String userId) {
         NoteModel noteModel = NoteModel.builder()
                 .topic(noteContainerModel.getTopic())
                 .text(noteContainerModel.getText())
                 .build();
-        //TODO: Add User control for ownership. Is notebook owner and user is same
+        BooleanResponse booleanResponse = noteBookClient.getIsUserOwner(noteContainerModel.getNoteBookId().toString());
+        if (!booleanResponse.getResult())
+            throw new UnauthorizedException(USER_DONT_HAVE_PERMISSION);
         noteModel = noteRepository.save(noteModel);
         noteContainerModel.setNoteId(noteModel.getId());
         noteContainerModel = noteContainerRepository.save(noteContainerModel);
@@ -48,11 +54,13 @@ public class NoteServiceImpl implements NoteService {
         return noteContainerModel;
     }
 
-    public NoteContainerModel updateNote(NoteContainerModel noteContainerModel) {
+    public NoteContainerModel updateNote(NoteContainerModel noteContainerModel, String userId) {
         UUID noteId = noteContainerModel.getId();
         NoteContainerModel oldNoteContainerModel = noteContainerRepository.findById(noteId)
                 .orElseThrow(() -> new ResourceNotFoundException("Note", "by id", noteId.toString()));
-        //TODO: Add User control for ownership. Is notebook owner and user is same
+        BooleanResponse booleanResponse = noteBookClient.getIsUserOwner(noteContainerModel.getNoteBookId().toString());
+        if (!booleanResponse.getResult())
+            throw new UnauthorizedException(USER_DONT_HAVE_PERMISSION);
         TypeMap<NoteContainerModel, NoteVersionModel> propertyMapper =
                 modelMapper.createTypeMap(NoteContainerModel.class, NoteVersionModel.class);
         propertyMapper.addMapping(NoteContainerModel::getId, NoteVersionModel::setNoteContainerId);
@@ -73,29 +81,37 @@ public class NoteServiceImpl implements NoteService {
         return noteContainerModel;
     }
 
-    public NoteContainerModel getNote(UUID id) {
+    public NoteContainerModel getNote(UUID id, String userId) {
         NoteContainerModel noteContainerModel = noteContainerRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Note", "by id", id.toString()));
         NoteModel noteModel = noteRepository.findById(noteContainerModel.getNoteId())
                 .orElseThrow(() -> new ResourceNotFoundException("Note", "by id", id.toString()));
-        //TODO: Add User control for ownership. Is notebook owner and user is same
+        BooleanResponse booleanResponse = noteBookClient.getIsUserOwner(noteContainerModel.getNoteBookId().toString());
+        if (!booleanResponse.getResult())
+            throw new UnauthorizedException(USER_DONT_HAVE_PERMISSION);
         noteContainerModel.setNoteId(noteModel.getId());
         noteContainerModel.setTopic(noteModel.getTopic());
         noteContainerModel.setText(noteModel.getText());
         return noteContainerModel;
     }
 
-    public List<NoteVersionModel> getNoteVersions(UUID id) {
+    public List<NoteVersionModel> getNoteVersions(UUID id, String userId) {
         List<NoteVersionModel> noteVersionModels = noteVersionRepository.findByNoteContainerIdEquals(id);
-        //TODO: Add User control for ownership. Is notebook owner and user is same
+        if (!noteVersionModels.isEmpty()) {
+            BooleanResponse booleanResponse = noteBookClient.getIsUserOwner(noteVersionModels.get(0).getNoteBookId().toString());
+            if (!booleanResponse.getResult())
+                throw new UnauthorizedException(USER_DONT_HAVE_PERMISSION);
+        }
         return noteVersionModels;
     }
 
 
-    public ApiResponse deleteNote(ByIdRequest byIdRequest) {
+    public ApiResponse deleteNote(ByIdRequest byIdRequest, String userId) {
         NoteContainerModel noteContainerModel = noteContainerRepository.findById(byIdRequest.getId())
                 .orElseThrow(() -> new ResourceNotFoundException("Note", "by id", byIdRequest.getId().toString()));
-        //TODO: Add User control for ownership. Is notebook owner and user is same
+        BooleanResponse booleanResponse = noteBookClient.getIsUserOwner(noteContainerModel.getNoteBookId().toString());
+        if (!booleanResponse.getResult())
+            throw new UnauthorizedException(USER_DONT_HAVE_PERMISSION);
         noteContainerRepository.delete(noteContainerModel);
         NoteModel noteModel = NoteModel.builder().id(noteContainerModel.getNoteId()).build();
         noteRepository.delete(noteModel);
